Divide publish tasks
